refactor: move /add_quote to sep files
